feat: refactor diagram store with new node and edge actions, introduce helper functions, and define DiagramState interface
