+++ frontend/features/diagram/store/diagramStore.types.ts
@@ -1,0 +1,41 @@
+import type { Connection, EdgeChange, NodeChange } from "@xyflow/react";
+import type { TableNode, RelationEdge, RelationshipType } from "../types/flow.types";
+import type { DbColumn } from "../types/db.types";
+
+export interface DiagramState {
+    nodes: TableNode[];
+    edges: RelationEdge[];
+    onNodesChange: (changes: NodeChange[]) => void;
+    onEdgesChange: (changes: EdgeChange[]) => void;
+    /**
+     * Creates an edge between two tables, auto-creating a FK column
+     * in the target table pointing at the source table's PK.
+     */
+    addEdgeWithType: (connection: Connection, fkName: string, type: RelationshipType) => void;
+    setEdgeRelationType: (edgeId: string, type: RelationshipType) => void;
+    /** Deletes the edge and removes the auto-created FK column from its node. */
+    deleteEdge: (edgeId: string) => void;
+    addTable: (name: string) => void;
+    updateColumn: (nodeId: string, column: DbColumn) => void;
+    addColumn: (nodeId: string, columnId?: string) => void;
+    removeColumn: (nodeId: string, columnId: string) => void;
+    renameTable: (nodeId: string, name: string) => void;
+    deleteTable: (nodeId: string) => void;
+    /** Creates a junction table for M:N and wires it with 1:N edges to both parent tables. */
+    createJunctionTable: (sourceNodeId: string, targetNodeId: string) => void;
+    /** Replaces the entire diagram state — used for loading saved or mock data. */
+    loadDiagram: (nodes: TableNode[], edges: RelationEdge[]) => void;
+    /** Toggles the handle side (left ↔ right) for all edges connected to a specific column. */
+    flipColumnHandleSide: (nodeId: string, columnId: string) => void;
+    /** Toggles the source or target handle of a single edge between left and right. */
+    flipEdgeEnd: (edgeId: string, end: "source" | "target") => void;
+    /**
+     * Re-targets a FK column to a different reference table:
+     * renames the column to {newTable}_id, updates its references,
+     * and swaps the connecting edge to start from the new table's PK.
+     */
+    retargetFkColumn: (nodeId: string, columnId: string, newRefTableId: string) => void;
+}
+
+/** Zustand setter typed against DiagramState. */
+export type SetState = (fn: (s: DiagramState) => Partial<DiagramState>) => void;
